C++: Split off stringstream.cpp.
