+++ cpp/ql/test/library-tests/dataflow/taint-tests/taint.expected
@@ -32,34 +32,34 @@
 | movableclass.cpp:55:8:55:9 | s2 | movableclass.cpp:52:23:52:28 | call to source |
 | movableclass.cpp:64:8:64:9 | s2 | movableclass.cpp:23:55:23:60 | call to source |
 | movableclass.cpp:65:11:65:11 | call to operator= | movableclass.cpp:65:13:65:18 | call to source |
-| string.cpp:29:7:29:7 | a | string.cpp:25:12:25:17 | call to source |
-| string.cpp:31:7:31:7 | c | string.cpp:27:16:27:21 | call to source |
-| string.cpp:33:9:33:13 | call to c_str | string.cpp:27:16:27:21 | call to source |
-| string.cpp:75:13:75:17 | call to c_str | string.cpp:14:10:14:15 | call to source |
-| string.cpp:79:13:79:17 | call to c_str | string.cpp:14:10:14:15 | call to source |
-| string.cpp:82:13:82:17 | call to c_str | string.cpp:14:10:14:15 | call to source |
-| string.cpp:92:7:92:8 | cs | string.cpp:87:19:87:24 | call to source |
-| string.cpp:93:7:93:8 | ss | string.cpp:87:19:87:24 | call to source |
-| string.cpp:106:7:106:8 | cs | string.cpp:98:19:98:24 | call to source |
-| string.cpp:107:7:107:8 | ss | string.cpp:98:19:98:24 | call to source |
-| string.cpp:129:8:129:9 | s1 | string.cpp:124:18:124:23 | call to source |
-| string.cpp:130:8:130:9 | s2 | string.cpp:125:20:125:25 | call to source |
-| string.cpp:131:8:131:9 | s3 | string.cpp:127:8:127:13 | call to source |
-| string.cpp:150:8:150:9 | s1 | string.cpp:146:32:146:37 | call to source |
-| string.cpp:151:8:151:9 | s2 | string.cpp:148:20:148:25 | call to source |
-| string.cpp:158:8:158:8 | c | string.cpp:156:16:156:21 | call to source |
-| string.cpp:166:8:166:8 | c | string.cpp:156:16:156:21 | call to source |
-| string.cpp:171:8:171:8 | c | string.cpp:169:28:169:33 | call to source |
-| string.cpp:181:11:181:11 | call to operator+ | string.cpp:178:18:178:23 | call to source |
-| string.cpp:182:11:182:11 | call to operator+ | string.cpp:178:18:178:23 | call to source |
-| string.cpp:183:11:183:11 | call to operator+ | string.cpp:178:18:178:23 | call to source |
-| string.cpp:186:11:186:11 | call to operator+ | string.cpp:186:13:186:18 | call to source |
-| string.cpp:195:8:195:9 | s5 | string.cpp:191:18:191:23 | call to source |
-| string.cpp:199:8:199:9 | s6 | string.cpp:191:18:191:23 | call to source |
-| string.cpp:204:8:204:9 | s7 | string.cpp:202:9:202:14 | call to source |
-| string.cpp:208:8:208:9 | s8 | string.cpp:191:18:191:23 | call to source |
-| string.cpp:213:8:213:9 | s9 | string.cpp:211:13:211:18 | call to source |
-| string.cpp:221:8:221:10 | s10 | string.cpp:218:12:218:26 | call to source |
+| string.cpp:28:7:28:7 | a | string.cpp:24:12:24:17 | call to source |
+| string.cpp:30:7:30:7 | c | string.cpp:26:16:26:21 | call to source |
+| string.cpp:32:9:32:13 | call to c_str | string.cpp:26:16:26:21 | call to source |
+| string.cpp:38:13:38:17 | call to c_str | string.cpp:14:10:14:15 | call to source |
+| string.cpp:42:13:42:17 | call to c_str | string.cpp:14:10:14:15 | call to source |
+| string.cpp:45:13:45:17 | call to c_str | string.cpp:14:10:14:15 | call to source |
+| string.cpp:55:7:55:8 | cs | string.cpp:50:19:50:24 | call to source |
+| string.cpp:56:7:56:8 | ss | string.cpp:50:19:50:24 | call to source |
+| string.cpp:69:7:69:8 | cs | string.cpp:61:19:61:24 | call to source |
+| string.cpp:70:7:70:8 | ss | string.cpp:61:19:61:24 | call to source |
+| string.cpp:92:8:92:9 | s1 | string.cpp:87:18:87:23 | call to source |
+| string.cpp:93:8:93:9 | s2 | string.cpp:88:20:88:25 | call to source |
+| string.cpp:94:8:94:9 | s3 | string.cpp:90:8:90:13 | call to source |
+| string.cpp:113:8:113:9 | s1 | string.cpp:109:32:109:37 | call to source |
+| string.cpp:114:8:114:9 | s2 | string.cpp:111:20:111:25 | call to source |
+| string.cpp:121:8:121:8 | c | string.cpp:119:16:119:21 | call to source |
+| string.cpp:129:8:129:8 | c | string.cpp:119:16:119:21 | call to source |
+| string.cpp:134:8:134:8 | c | string.cpp:132:28:132:33 | call to source |
+| string.cpp:144:11:144:11 | call to operator+ | string.cpp:141:18:141:23 | call to source |
+| string.cpp:145:11:145:11 | call to operator+ | string.cpp:141:18:141:23 | call to source |
+| string.cpp:146:11:146:11 | call to operator+ | string.cpp:141:18:141:23 | call to source |
+| string.cpp:149:11:149:11 | call to operator+ | string.cpp:149:13:149:18 | call to source |
+| string.cpp:158:8:158:9 | s5 | string.cpp:154:18:154:23 | call to source |
+| string.cpp:162:8:162:9 | s6 | string.cpp:154:18:154:23 | call to source |
+| string.cpp:167:8:167:9 | s7 | string.cpp:165:9:165:14 | call to source |
+| string.cpp:171:8:171:9 | s8 | string.cpp:154:18:154:23 | call to source |
+| string.cpp:176:8:176:9 | s9 | string.cpp:174:13:174:18 | call to source |
+| string.cpp:184:8:184:10 | s10 | string.cpp:181:12:181:26 | call to source |
 | structlikeclass.cpp:35:8:35:9 | s1 | structlikeclass.cpp:29:22:29:27 | call to source |
 | structlikeclass.cpp:36:8:36:9 | s2 | structlikeclass.cpp:30:24:30:29 | call to source |
 | structlikeclass.cpp:37:8:37:9 | s3 | structlikeclass.cpp:29:22:29:27 | call to source |
