+++ cpp/ql/test/library-tests/dataflow/taint-tests/test_diff.expected
@@ -30,33 +30,33 @@
 | movableclass.cpp:55:8:55:9 | movableclass.cpp:52:23:52:28 | AST only |
 | movableclass.cpp:64:8:64:9 | movableclass.cpp:23:55:23:60 | AST only |
 | movableclass.cpp:65:11:65:11 | movableclass.cpp:65:13:65:18 | AST only |
-| string.cpp:31:7:31:7 | string.cpp:27:16:27:21 | AST only |
-| string.cpp:33:9:33:13 | string.cpp:27:16:27:21 | AST only |
-| string.cpp:75:13:75:17 | string.cpp:14:10:14:15 | AST only |
-| string.cpp:79:13:79:17 | string.cpp:14:10:14:15 | AST only |
-| string.cpp:82:13:82:17 | string.cpp:14:10:14:15 | AST only |
-| string.cpp:92:7:92:8 | string.cpp:87:19:87:26 | IR only |
-| string.cpp:93:7:93:8 | string.cpp:87:19:87:24 | AST only |
-| string.cpp:106:7:106:8 | string.cpp:98:19:98:24 | AST only |
-| string.cpp:107:7:107:8 | string.cpp:98:19:98:24 | AST only |
-| string.cpp:129:8:129:9 | string.cpp:124:18:124:23 | AST only |
-| string.cpp:130:8:130:9 | string.cpp:125:20:125:25 | AST only |
-| string.cpp:131:8:131:9 | string.cpp:127:8:127:13 | AST only |
-| string.cpp:150:8:150:9 | string.cpp:146:32:146:37 | AST only |
-| string.cpp:151:8:151:9 | string.cpp:148:20:148:25 | AST only |
-| string.cpp:158:8:158:8 | string.cpp:156:16:156:21 | AST only |
-| string.cpp:166:8:166:8 | string.cpp:156:16:156:21 | AST only |
-| string.cpp:171:8:171:8 | string.cpp:169:28:169:33 | AST only |
-| string.cpp:181:11:181:11 | string.cpp:178:18:178:23 | AST only |
-| string.cpp:182:11:182:11 | string.cpp:178:18:178:23 | AST only |
-| string.cpp:183:11:183:11 | string.cpp:178:18:178:23 | AST only |
-| string.cpp:186:11:186:11 | string.cpp:186:13:186:18 | AST only |
-| string.cpp:195:8:195:9 | string.cpp:191:18:191:23 | AST only |
-| string.cpp:199:8:199:9 | string.cpp:191:18:191:23 | AST only |
-| string.cpp:204:8:204:9 | string.cpp:202:9:202:14 | AST only |
-| string.cpp:208:8:208:9 | string.cpp:191:18:191:23 | AST only |
-| string.cpp:213:8:213:9 | string.cpp:211:13:211:18 | AST only |
-| string.cpp:221:8:221:10 | string.cpp:218:12:218:26 | AST only |
+| string.cpp:30:7:30:7 | string.cpp:26:16:26:21 | AST only |
+| string.cpp:32:9:32:13 | string.cpp:26:16:26:21 | AST only |
+| string.cpp:38:13:38:17 | string.cpp:14:10:14:15 | AST only |
+| string.cpp:42:13:42:17 | string.cpp:14:10:14:15 | AST only |
+| string.cpp:45:13:45:17 | string.cpp:14:10:14:15 | AST only |
+| string.cpp:55:7:55:8 | string.cpp:50:19:50:26 | IR only |
+| string.cpp:56:7:56:8 | string.cpp:50:19:50:24 | AST only |
+| string.cpp:69:7:69:8 | string.cpp:61:19:61:24 | AST only |
+| string.cpp:70:7:70:8 | string.cpp:61:19:61:24 | AST only |
+| string.cpp:92:8:92:9 | string.cpp:87:18:87:23 | AST only |
+| string.cpp:93:8:93:9 | string.cpp:88:20:88:25 | AST only |
+| string.cpp:94:8:94:9 | string.cpp:90:8:90:13 | AST only |
+| string.cpp:113:8:113:9 | string.cpp:109:32:109:37 | AST only |
+| string.cpp:114:8:114:9 | string.cpp:111:20:111:25 | AST only |
+| string.cpp:121:8:121:8 | string.cpp:119:16:119:21 | AST only |
+| string.cpp:129:8:129:8 | string.cpp:119:16:119:21 | AST only |
+| string.cpp:134:8:134:8 | string.cpp:132:28:132:33 | AST only |
+| string.cpp:144:11:144:11 | string.cpp:141:18:141:23 | AST only |
+| string.cpp:145:11:145:11 | string.cpp:141:18:141:23 | AST only |
+| string.cpp:146:11:146:11 | string.cpp:141:18:141:23 | AST only |
+| string.cpp:149:11:149:11 | string.cpp:149:13:149:18 | AST only |
+| string.cpp:158:8:158:9 | string.cpp:154:18:154:23 | AST only |
+| string.cpp:162:8:162:9 | string.cpp:154:18:154:23 | AST only |
+| string.cpp:167:8:167:9 | string.cpp:165:9:165:14 | AST only |
+| string.cpp:171:8:171:9 | string.cpp:154:18:154:23 | AST only |
+| string.cpp:176:8:176:9 | string.cpp:174:13:174:18 | AST only |
+| string.cpp:184:8:184:10 | string.cpp:181:12:181:26 | AST only |
 | structlikeclass.cpp:35:8:35:9 | structlikeclass.cpp:29:22:29:27 | AST only |
 | structlikeclass.cpp:36:8:36:9 | structlikeclass.cpp:30:24:30:29 | AST only |
 | structlikeclass.cpp:37:8:37:9 | structlikeclass.cpp:29:22:29:27 | AST only |
